2017.8.15更新1
速度125
增加偏移过灯 距离7.6
非常稳定

+++ new/Project_Headers/steer.h
@@ -21,7 +21,7 @@ extern byte cycle_i,cycle_j,turnleft;
 extern double target_offset,last_offset,Steer_kp;
 extern WORD Steer_PWM[4];
 
-void SET_steer(unsigned int steer);
+void SET_steer(double steer);
 void steer_error_left(void);
 void steer_error_right(void);
 extern void Steer_PDSet(void);
